% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.pattern.r
\name{change.pattern}
\alias{change.pattern}
\title{Pattern Changes}
\usage{
change.pattern(changes, pattern, tolerance)
}
\arguments{
\item{changes}{A list as returned by \code{\link{genelines}} or \code{\link{genelinesLog2FC}}.}

\item{pattern}{A named list of the desired values for each comparison. Names should correspond to the column names in \emph{changes$slopes}.}

\item{tolerance}{A single number corresponding to how much above/below each value in \emph{pattern} a gene may be and still be returned.}
}
\value{
A list of two data frames: the first contains the mean value of each treatment group for the scaled and centred counts of each gene, and the second contains the differences between pairs of treatment levels.
}
\description{
Find genes with a particular expression pattern
}
\details{
The differences between pairs of treatment levels is equivalent to the slope of a line drawn between the treatment means. \emph{change.pattern} allows the user to define a set of numerical relationships between treatments and find which genes follow this pattern. The amount of deviation allowed from the specified pattern is determined by \emph{tolerance}.Note that if \emph{changes was produced by  \code{\link{genelinesLog2FC}}}, the slopes correspond to log2 fold changes.
}
\examples{
genes<-matrix(round(rnorm(n=600, mean=100, sd=10), 0), ncol = 6, dimnames = list(paste0("g", seq(1:100))))
treats<-as.factor(c(rep("control",3), rep("treated",3)))
changes<-genelines(genes, treats)
change.pattern(changes, pattern = list(C1_2 = 0.5), tolerance = 0.2)

}
\seealso{
\code{\link{change.direction}}

\code{\link{change.similar}}
}
