% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.directions.r
\name{combined.directions}
\alias{combined.directions}
\title{Direction Changes}
\usage{
combined.directions(changes, label.plots=NULL, ...)
}
\arguments{
\item{changes}{A list as returned by \code{\link{genelines}} or \code{\link{genelinesLog2FC}}.}

\item{label.plots}{A single keyword for position (as supplied to \code{\link[graphics]{legend}}) for labelling the plots with the number of genes in each group. Defaults to NULL.}

\item{...}{Other parameters to be passed to \code{\link{change.direction}}.}
}
\value{
A list of eight, containing the genes that follow the patterns flat-up,   flat-down, up-up, up-flat, up-down,  down-up, down-flat, down-down. Within each item are two data frames: the first contains the mean value of each treatment group for each gene, and the second contains the differences between pairs of treatment levels.
}
\description{
Find and plot genes with all possible expression patterns for two comparisons
}
\details{
The differences between pairs of treatment levels is equivalent to the slope of a line drawn between the treatment means. \code{\link{change.direction}} allows the user to define a single set of relationships between treatments and find which genes follow this pattern. The amount of deviation allowed from the specified pattern is determined by \emph{tolerance}.
}
\examples{
genes<-matrix(round(rnorm(n=900, mean=100, sd=10), 0),
ncol = 9, dimnames = list(paste0("g", seq(1:100))))
treats<-as.factor(c(rep("control",3), rep("treat1",3), rep("treat2",3)))
changes<-genelines(genes, treats)
combined.directions(changes, tolerance = 0.2, label.plots="center")
}
\seealso{
\code{\link{genelines}}

\code{\link{genelinesLog2FC}}

\code{\link{change.direction}}
}
