% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneclust.r
\name{geneclust}
\alias{geneclust}
\title{geneclust}
\usage{
geneclust(changes, k, cor.method = "pearson", clust.method = "complete")
}
\arguments{
\item{changes}{A list as returned by \code{\link{genelines}} or \code{\link{genelinesLog2FC}}.}

\item{k}{The desired number of clusters.}

\item{cor.method}{The correlation method for the distance matrix, as passed to \code{\link[stats]{cor}}.Defaults to "pearson".}

\item{clust.method}{The clustering method, as passed to \code{\link[stats]{hclust}}.Defaults to "complete".}
}
\value{
A list of cluster membership for each gene and a plot of each cluster.
}
\description{
Perform cluster analysis on genelines output.
}
\examples{
genes<-matrix(round(rnorm(n=600, mean=100, sd=10), 0),
 ncol = 6, dimnames = list(paste0("g", seq(1:100))))
treats<-as.factor(c(rep("control",3), rep("treated",3)))
changes<-genelines(genes, treats)
geneclust(changes, k=3)
}
\seealso{
\code{\link{genelines}}

\code{\link{genelinesLog2FC}}
}
