% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genelines.r
\name{genelines}
\alias{genelines}
\title{Genelines}
\usage{
genelines(obj, condition, comparisons = NULL)
}
\arguments{
\item{obj}{A data frame or matrix of normalised gene counts; genes are rows, samples are columns. Row names should correspond to Gene IDs.}

\item{condition}{A list of the treatment group for each column in obj.}

\item{comparisons}{A list of desired comparisons, in the format \emph{list("1-2", "2-3", "1-3")}. This would be equivalent to treatment 2 - treatment 1, treatment 3 - treatment 2, and treatment 3 - treatment 1. If NULL (the default), comparisions are performed between successive treatment levels in \emph{condition}. Note that treatment names must NOT contain dashes (-), as this is used to separate treatment pairs.}
}
\value{
A list of two data frames: the first (means) contains the mean value of each treatment group for the scaled and centred counts of each gene, and the second (slopes) contains the differences between pairs of treatment levels.
}
\description{
Calculate changes between treatments
}
\details{
Note that although features are referred to as genes for convenience, this function is best suited for non-gene data. If you have RNAseq data you are advised to use \code{\link{genelinesLog2FC}}.
Treatment levels are compared in sequential pairs, i.e. treatment 2 - treatment 1, treatment 3 - treatment 2, unless otherwise specified by *comparisons*. This is equivalent to the slope of a line drawn between the treatment means.
}
\examples{
genes<-matrix(round(rnorm(n=600, mean=100, sd=10), 0),
ncol = 6, dimnames = list(paste0("g", seq(1:100))))
treats<-as.factor(c(rep("control",3), rep("treated",3)))
genelines(genes, treats)

}
\seealso{
\code{\link{graph.genelines}}
}
