% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genelinesLog2FC.r
\name{genelinesLog2FC}
\alias{genelinesLog2FC}
\title{Genelines for log2 fold change data}
\usage{
genelinesLog2FC(log2fc, condition, sig.level=0.05, FC, degs="inclusive")
}
\arguments{
\item{data}{A list of DESeq results objects, in the order you wish to make the comparisons.}

\item{condition}{A vector of the treatment group levels, in the same order they appear in data. Note that treatment names must NOT contain dashes (-), as this is used to separate treatment pairs.}

\item{sig.level}{The significance cutoff used to decide a differentially expressed gene. Defaults to 0.05; can be removed by setting to NULL.}

\item{FC}{The fold change cutoff used to decide a differentially expressed gene. Defaults to 0 (no cutoff).}

\item{degs}{Whether to return genes differentially expressed in ANY comparison ("inclusive") or ALL comparisons ("exclusive"). Defaults to inclusive.}
}
\value{
A list of two data frames: the first (means) contains the mean value of each treatment group relative to the first treatment group, and the second (slopes) contains the log2 fold changes between pairs of treatment levels.
}
\description{
Calculate expression changes between treatments
}
\details{
Treatment levels are compared in the order supplied by *data*, i.e. treatment 2 - treatment 1, treatment 3 - treatment 2. The 'means' are for compatibility with other *genelines* functions only. All genes are assigned a value of 0 for treatment 1. Genes are only included in the results if they meet the *sig.level* and *FC* criteria.
}
\examples{
genes<-list(deData1=data.frame(log2FoldChange=round(rnorm(n=100, mean=100, sd=10), 0)), deData2=data.frame(log2FoldChange=round(rnorm(n=100, mean=100, sd=10), 0)))
rownames(genes$deData1)<-paste0("g", seq(1:100))
rownames(genes$deData2)<-paste0("g", seq(1:100))
genelinesLog2FC(genes, c("control", "treat1", "treat2"))

}
\seealso{
\code{\link{graph.genelines}}
}
