% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.genelines.r
\name{graph.genelines}
\alias{graph.genelines}
\title{graph.genelines}
\usage{
graph.genelines(changes, ...)
}
\arguments{
\item{changes}{A list as returned by \code{\link{genelines}} or \code{\link{genelinesLog2FC}}.}

\item{...}{Additional parameters to be passed on to \code{\link{lines}}}
}
\value{
A plot showing the change in mean value of each treatment group for the scaled and centred counts of each gene.
}
\description{
Plotting function for genelines
}
\examples{
genes<-matrix(round(rnorm(n=600, mean=100, sd=10), 0),
ncol = 6, dimnames = list(paste0("g", seq(1:100))))
treats<-as.factor(c(rep("control",3), rep("treated",3)))
changes<-genelines(genes, treats)
graph.genelines(changes)
}
\seealso{
\code{\link{genelines}}

\code{\link{genelinesLog2FC}}
}
